feat(hello-work-software-jobs): /jobsページをslot構成（@search/@detail）にリファクタ＆レイアウト刷新
- Next.jsのslot機能を用いた2カラム分割レイアウトに移行
- ルーティング・ディレクトリ構成の整理
- 関連ファイルの移動・新規作成・削除
