+++ apps/hello-work-job-searcher/src/app/jobs/@detail/[jobNumber]/layout.tsx
@@ -1,0 +1,11 @@
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <div>
+      <div>{children}</div>
+    </div>
+  );
+}
