+++ apps/hello-work-job-searcher/src/app/jobs/@detail/[jobNumber]/loading.tsx
@@ -1,4 +1,4 @@
-import Loading from "../loading";
+import Loading from "../../loading";
 
 export default function Page() {
   // Or a custom loading skeleton component
