+++ apps/hello-work-job-searcher/src/app/jobs/@detail/[jobNumber]/page.module.css
