+++ apps/hello-work-job-searcher/src/app/jobs/@detail/default.tsx
@@ -1,0 +1,7 @@
+export default function DefaultDetail() {
+    return (
+        <div>
+            <p>求人を選択してください</p>
+        </div>
+    )
+}
